Common to multiStepHelpers refactor
