+++ extensions/vscode/src/multiStepInputs/common.ts
@@ -5,7 +5,6 @@ import {
   Credential,
   SnowflakeConnection,
   ServerType,
-  ProductType,
   ProductName,
   ProductDescription,
 } from "../api";
@@ -31,8 +30,10 @@ import {
 } from "../api/types/connectCloud";
 import axios from "axios";
 import { showProgress } from "../utils/progress";
-
-export const createNewCredentialLabel = "Create a New Credential";
+import {
+  isConnectCloud,
+  createNewCredentialLabel,
+} from "../utils/multiStepHelpers";
 
 // Search for the first credential that includes
 // the targetURL.
@@ -46,59 +47,6 @@ export function findExistingCredentialByURL(
     return newURL.includes(existing);
   });
 }
-
-export const isConnect = (serverType: ServerType) => {
-  return serverType === ServerType.CONNECT;
-};
-
-export const isConnectCloud = (serverType: ServerType) => {
-  return serverType === ServerType.CONNECT_CLOUD;
-};
-
-export const isSnowflake = (serverType: ServerType) => {
-  return serverType === ServerType.SNOWFLAKE;
-};
-
-export const isConnectProduct = (productType: ProductType) => {
-  return productType === ProductType.CONNECT;
-};
-
-export const isConnectCloudProduct = (productType: ProductType) => {
-  return productType === ProductType.CONNECT_CLOUD;
-};
-
-export const getProductType = (serverType: ServerType): ProductType => {
-  switch (serverType) {
-    case ServerType.CONNECT:
-      return ProductType.CONNECT;
-    case ServerType.SNOWFLAKE:
-      return ProductType.CONNECT;
-    case ServerType.CONNECT_CLOUD:
-      return ProductType.CONNECT_CLOUD;
-  }
-};
-
-export const getProductName = (productType: ProductType) => {
-  switch (productType) {
-    case ProductType.CONNECT:
-      return ProductName.CONNECT;
-    case ProductType.CONNECT_CLOUD:
-      return ProductName.CONNECT_CLOUD;
-  }
-};
-
-export const getServerType = (productName: ProductName) => {
-  switch (productName) {
-    case ProductName.CONNECT:
-      return ServerType.CONNECT;
-    case ProductName.CONNECT_CLOUD:
-      return ServerType.CONNECT_CLOUD;
-  }
-};
-
-export const getPublishableAccounts = (accounts: ConnectCloudAccount[]) => {
-  return accounts.filter((a) => a.permissionToPublish);
-};
 
 // List of all available platforms
 export const platformList: QuickPickItem[] = [
