+++ extensions/vscode/webviews/homeView/vite.config.ts
@@ -20,7 +20,6 @@ export default defineConfig({
   },
   build: {
     rollupOptions: {
-      external: ["vscode"], // Exclude 'vscode' module
       output: {
         entryFileNames: `[name].js`,
         chunkFileNames: `[name].js`,
